+++ packages/nimble-components/src/card-button/tests/card-button.spec.ts
@@ -1,0 +1,9 @@
+import { CardButton } from '..';
+
+describe('Card Button', () => {
+    it('can construct an element instance', () => {
+        expect(document.createElement('nimble-card-button')).toBeInstanceOf(
+            CardButton
+        );
+    });
+});