+++ packages/nimble-components/src/dialog/template.ts
@@ -1,0 +1,15 @@
+import { html, ref } from '@microsoft/fast-element';
+import type { Dialog } from '.';
+
+export const template = html<Dialog>`
+    <template>
+        <dialog
+            ${ref('dialogElement')}
+            role="alertdialog"
+            aria-label="${x => x.ariaLabel}"
+            @cancel="${(x, c) => x.cancelHandler(c.event)}"
+        >
+            <slot></slot>
+        </dialog>
+    </template>
+`;