+++ packages/nimble-components/src/icon-base/types.ts
@@ -6,6 +6,7 @@ export const IconStatus = {
     fail: 'fail',
     warning: 'warning',
     pass: 'pass',
-    regular: 'regular'
+    regular: 'regular',
+    information: 'information'
 } as const;
 export type IconStatus = typeof IconStatus[keyof typeof IconStatus];