+++ packages/nimble-components/src/text-field/styles.ts
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/indent */
 import { css } from '@microsoft/fast-element';
 import { display } from '@microsoft/fast-foundation';
 
@@ -9,7 +10,6 @@ import {
     bodyDisabledFontColor,
     controlHeight,
     failColor,
-    fillSelectedRgbPartialColor,
     labelHeight,
     smallDelay,
     controlLabelFont,
@@ -152,11 +152,6 @@ export const styles = css`
         text-overflow: clip;
     }
 
-    .control::selection {
-        color: ${controlLabelFontColor};
-        background: rgba(${fillSelectedRgbPartialColor}, 0.3);
-    }
-
     .control::placeholder {
         color: ${controlLabelFontColor};
     }
@@ -207,18 +202,18 @@ export const styles = css`
         ${controlHeight.cssCustomProperty}: 24px;
     }
 `.withBehaviors(
-                appearanceBehavior(
-                    TextFieldAppearance.underline,
-                    css`
+    appearanceBehavior(
+        TextFieldAppearance.underline,
+        css`
             .root {
                 border-bottom-width: ${borderWidth};
                 padding-bottom: 0;
             }
         `
-                ),
-                appearanceBehavior(
-                    TextFieldAppearance.block,
-                    css`
+    ),
+    appearanceBehavior(
+        TextFieldAppearance.block,
+        css`
             .root {
                 background-color: rgba(${borderRgbPartialColor}, 0.1);
             }
@@ -247,19 +242,19 @@ export const styles = css`
                 background-color: rgba(${borderRgbPartialColor}, 0.07);
             }
         `
-                ),
-                appearanceBehavior(
-                    TextFieldAppearance.outline,
-                    css`
+    ),
+    appearanceBehavior(
+        TextFieldAppearance.outline,
+        css`
             .root {
                 border-width: ${borderWidth};
                 padding: 0;
             }
         `
-                ),
-                appearanceBehavior(
-                    TextFieldAppearance.frameless,
-                    css`
+    ),
+    appearanceBehavior(
+        TextFieldAppearance.frameless,
+        css`
             .control {
                 padding-left: ${borderWidth};
                 padding-right: ${borderWidth};
@@ -269,21 +264,21 @@ export const styles = css`
                 border-color: transparent;
             }
         `
-                ),
-                themeBehavior(
-                    css`
-            ${'' /* Light theme */}
+    ),
+    themeBehavior(
+        Theme.light,
+        css`
             .control::-ms-reveal {
                 filter: invert(0%);
             }
-        `,
-                    css`
-            ${'' /* Dark theme */}
+        `
+    ),
+    themeBehavior(
+        [Theme.dark, Theme.color],
+        css`
             .control::-ms-reveal {
                 filter: invert(100%);
             }
-        `,
-                    // Color theme
-                    Theme.dark
-                )
-            );
+        `
+    )
+);