+++ packages/nimble-components/src/theme-provider/design-tokens-static.ts
@@ -1,0 +1,6 @@
+import { Black } from '@ni/nimble-tokens/dist/styledictionary/js/tokens';
+import { hexToRgbaCssColor } from '../utilities/style/colors';
+
+export const modalBackdropColorThemeLight = hexToRgbaCssColor(Black, 0.3);
+export const modalBackdropColorThemeDark = hexToRgbaCssColor(Black, 0.6);
+export const modalBackdropColorThemeColor = hexToRgbaCssColor(Black, 0.6);